+++ backend/parking-backend/src/main/java/ru/app/parking_backend/service/ParkingSpaceService.java
@@ -1,7 +1,10 @@
 package ru.app.parking_backend.service;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.app.parking_backend.entity.Client;
 import ru.app.parking_backend.entity.ParkingSpace;
+import ru.app.parking_backend.exception.ResourceNotFoundException;
 import ru.app.parking_backend.repository.ParkingSpaceRepository;
 
 import java.util.List;
@@ -17,15 +20,23 @@ public class ParkingSpaceService {
         return repository.findAll();
     }
 
-    public Optional<ParkingSpace> findById(Integer id) {
-        return repository.findById(id);
+    public ParkingSpace findById(Integer id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Парковочное место с id " + id + " не найдено"));
     }
 
     public ParkingSpace save(ParkingSpace parkingSpace) {
+        if (parkingSpace.id() != null && !repository.existById(parkingSpace.id())) {
+            throw new ResourceNotFoundException("Парковочное место с id " + parkingSpace.id() + " не найдено");
+        }
         return repository.save(parkingSpace);
     }
 
     public void delete(Integer id) {
+        if (!repository.existById(id)) {
+            throw new ResourceNotFoundException("Парковочное место с id " + id + " не найдено");
+        }
         repository.delete(id);
     }
 }
+
